requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.0
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
basicsr==1.4.2
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.29.3
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.3.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gradio==3.16.2
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib_metadata==7.1.0
invisible-watermark==0.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.4
lightning-utilities==0.11.2
linkify-it-py==2.0.3
lit==18.1.6
lmdb==1.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pathtools==0.1.2
pi==0.1.2
pillow==10.2.0
platformdirs==4.2.2
prettytable==3.6.0
protobuf==3.20.1
psutil==5.9.8
pyarrow==16.1.0
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.1.2
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.5.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qudida==0.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.28.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.2.7
scikit-image==0.20.0
scikit-learn==1.5.0
scipy==1.13.1
semver==3.0.2
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
sympy==1.12
tb-nightly==2.17.0a20240520
tensorboard==2.16.2
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==0.6.12
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.6.0
torchvision==0.15.2
tornado==6.4
tqdm==4.66.4
transformers==4.19.2
triton==2.0.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==1.26.13
uvicorn==0.29.0
uvloop==0.19.0
validators==0.28.1
wandb==0.15.0
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==12.0
Werkzeug==3.0.3
xformers==0.0.22
yapf==0.32.0
yarl==1.9.4
zipp==3.18.2
